move handleThreshold to utils
